backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
python-multipart
pytest
httpx